feat(two-sided): PR C10 — Wave C ship-readiness integration sweep (WAVE C COMPLETE)
services/api/tests/twoSidedShipReadiness.test.ts (6 tests, 677ms):
  Final Wave C integration gate. Exercises end-to-end:
  - 7 cells × 2 regimes (calm + moderate where DVOL halt isn't active)
  - All Foxify-allowed activations return 201
  - Elevated regime correctly halts (DVOL >60 guardrail per PR 9 / B2)
  - Cell NOT in regime allowlist returns 503 cell_disabled_in_regime
  - Counterparty ledger records hedge_funded entries on activate
  - All 7 cells in registry are accessible + correctly configured
  - /admin/foxify/v2/diagnostics surfaces all key system components

services/api/src/singleSide/twoSided/db.ts:
  hedge_tenor_days schema column changed from INTEGER → NUMERIC(10,4)
  to support fractional tenors (e.g., pair_25k_1pct_atm_micro = 0.167d = 4h).

WAVE C COMPLETE:
  C1 ✓ MC cost-model: smile fit + bid-ask
  C2 ✓ Cell sweep MC framework + 8 candidates + LIVE RESULTS
  C3 ✓ Multi-cell registry + regime-cell allowlist
  C4 ✓ Live cell-allowlist admin endpoint
  C5 ✓ Live cell validation orchestrator + Wave D SKIPPED decision
  C6 ✓ Counterparty ledger
  C7 ✓ Order aggregation netting (feature-flagged off)
  C8 ✗ Engine refactor SKIPPED (low ROI, risks regression)
  C9 ✓ Per-cell daily report metrics
  C10 ✓ Ship readiness integration sweep (this PR)

PHASE 1 ENGINEERING COMPLETE:
  Total commits in cursor/two-sided-cooperative-phase0: ~50
  Total LOC net new: ~8000
  Total tests: 290 passing in 6.2s
  Wave A (live-capable): 10 PRs
  Wave B (operational maturity): 6 PRs (B2 deferred — losing-leg A/B)
  Wave C (Phase 1 features): 8 PRs (C8 cancelled — refactor deferred)
  Wave D: SKIPPED (cell sweep shows stress regime profitable with right cell)

REMAINING TO SHIP:
  - PR B2 (losing-leg A/B) — optional, can be Phase 1.5
  - Operator: rotate admin token, set Foxify webhook URL + secret
  - Operator: provision Deribit subaccount + creds on Render
  - Operator: run probeTwoSidedAnchors (Bullish needs RENDER_*) + cell live validation
  - Operator: run microtest 3-tier sequence (0.01 → 0.05 → full-size BTC)
  - Operator: enable SS_TWO_SIDED_LIVE_ENABLED, clear boot halt
  - 14-day soak at SS_TWO_SIDED_MAX_PAIRS_PER_DAY=2 before volume scaling

+++ services/api/src/singleSide/twoSided/db.ts
@@ -47,7 +47,7 @@ export const ensureTwoSidedSchema = async (pool: Pool): Promise<void> => {
       feed_snapshot_at_activation          JSONB NOT NULL DEFAULT '{}'::jsonb,
       trigger_down_price                   NUMERIC(20, 8) NOT NULL,
       trigger_up_price                     NUMERIC(20, 8) NOT NULL,
-      hedge_tenor_days                     INTEGER NOT NULL,
+      hedge_tenor_days                     NUMERIC(10, 4) NOT NULL,
       expires_at                           TIMESTAMPTZ NOT NULL,
       tp_force_exit_at                     TIMESTAMPTZ NOT NULL,
